+++ crates/history/src/browser.rs
@@ -1,23 +1,17 @@
-use std::any::Any;
-use std::borrow::Cow;
-use std::cell::RefCell;
-use std::fmt;
-use std::rc::Rc;
+use std::{any::Any, borrow::Cow, cell::RefCell, fmt, rc::Rc};
 
 use gloo_events::EventListener;
 use gloo_utils::window;
-#[cfg(feature = "query")]
-use serde::Serialize;
 use wasm_bindgen::{JsValue, UnwrapThrowExt};
 use web_sys::Url;
 
-#[cfg(feature = "query")]
-use crate::error::HistoryResult;
 use crate::history::History;
 use crate::listener::HistoryListener;
 use crate::location::Location;
 use crate::state::{HistoryState, StateMap};
 use crate::utils::WeakCallback;
+#[cfg(feature = "query")]
+use crate::{error::HistoryResult, query::ToQuery};
 
 /// A [`History`] that is implemented with [`web_sys::History`] that provides native browser
 /// history and state access.
@@ -109,12 +103,16 @@ impl History for BrowserHistory {
     }
 
     #[cfg(feature = "query")]
-    fn push_with_query<'a, Q>(&self, route: impl Into<Cow<'a, str>>, query: Q) -> HistoryResult<()>
+    fn push_with_query<'a, Q>(
+        &self,
+        route: impl Into<Cow<'a, str>>,
+        query: Q,
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize,
+        Q: ToQuery,
     {
         let route = route.into();
-        let query = serde_urlencoded::to_string(query)?;
+        let query = query.to_query()?;
 
         let url = Self::combine_url(&route, &query);
 
@@ -131,12 +129,12 @@ impl History for BrowserHistory {
         &self,
         route: impl Into<Cow<'a, str>>,
         query: Q,
-    ) -> HistoryResult<()>
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize,
+        Q: ToQuery,
     {
         let route = route.into();
-        let query = serde_urlencoded::to_string(query)?;
+        let query = query.to_query()?;
 
         let url = Self::combine_url(&route, &query);
 
@@ -154,9 +152,9 @@ impl History for BrowserHistory {
         route: impl Into<Cow<'a, str>>,
         query: Q,
         state: T,
-    ) -> HistoryResult<()>
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize,
+        Q: ToQuery,
         T: 'static,
     {
         let (id, history_state) = Self::create_history_state();
@@ -165,7 +163,7 @@ impl History for BrowserHistory {
         states.insert(id, Rc::new(state) as Rc<dyn Any>);
 
         let route = route.into();
-        let query = serde_urlencoded::to_string(query)?;
+        let query = query.to_query()?;
 
         let url = Self::combine_url(&route, &query);
 
@@ -184,9 +182,9 @@ impl History for BrowserHistory {
         route: impl Into<Cow<'a, str>>,
         query: Q,
         state: T,
-    ) -> HistoryResult<()>
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize,
+        Q: ToQuery,
         T: 'static,
     {
         let (id, history_state) = Self::create_history_state();
@@ -195,7 +193,7 @@ impl History for BrowserHistory {
         states.insert(id, Rc::new(state) as Rc<dyn Any>);
 
         let route = route.into();
-        let query = serde_urlencoded::to_string(query)?;
+        let query = query.to_query()?;
 
         let url = Self::combine_url(&route, &query);