+++ crates/history/src/history.rs
@@ -1,12 +1,9 @@
 use std::borrow::Cow;
 
-#[cfg(feature = "query")]
-use serde::Serialize;
-
-#[cfg(feature = "query")]
-use crate::error::HistoryResult;
 use crate::listener::HistoryListener;
 use crate::location::Location;
+#[cfg(feature = "query")]
+use crate::{error::HistoryResult, query::ToQuery};
 
 /// A trait to provide [`History`] access.
 ///
@@ -56,9 +53,13 @@ pub trait History: Clone + PartialEq {
 
     /// Same as `.push()` but affix the queries to the end of the route.
     #[cfg(feature = "query")]
-    fn push_with_query<'a, Q>(&self, route: impl Into<Cow<'a, str>>, query: Q) -> HistoryResult<()>
+    fn push_with_query<'a, Q>(
+        &self,
+        route: impl Into<Cow<'a, str>>,
+        query: Q,
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize;
+        Q: ToQuery;
 
     /// Same as `.replace()` but affix the queries to the end of the route.
     #[cfg(feature = "query")]
@@ -66,9 +67,9 @@ pub trait History: Clone + PartialEq {
         &self,
         route: impl Into<Cow<'a, str>>,
         query: Q,
-    ) -> HistoryResult<()>
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize;
+        Q: ToQuery;
 
     /// Same as `.push_with_state()` but affix the queries to the end of the route.
     #[cfg(feature = "query")]
@@ -77,9 +78,9 @@ pub trait History: Clone + PartialEq {
         route: impl Into<Cow<'a, str>>,
         query: Q,
         state: T,
-    ) -> HistoryResult<()>
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize,
+        Q: ToQuery,
         T: 'static;
 
     /// Same as `.replace_with_state()` but affix the queries to the end of the route.
@@ -89,9 +90,9 @@ pub trait History: Clone + PartialEq {
         route: impl Into<Cow<'a, str>>,
         query: Q,
         state: T,
-    ) -> HistoryResult<()>
+    ) -> HistoryResult<(), Q::Error>
     where
-        Q: Serialize,
+        Q: ToQuery,
         T: 'static;
 
     /// Creates a Listener that will be notified when current state changes.