+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Next Version
 
 - Migrate to Edition 2021 and Apply MSRV in Cargo.toml (#360)
+- Introduces the `FromQuery` and `ToQuery` traits to allow for customizing
+  how query strings are encoded and decoded in `gloo_history`. (#364)
 
 ### Version "0.2.3"